requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0
faiss-cpu==1.7.4
numpy>=1.24.3
pandas>=2.0.3
python-multipart==0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==0.8.11
pillow>=10.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.1
requests>=2.31.0
streamlit>=1.31.0
